chore(core): Add `ingest_timestamp` to `EventMetadata`
We want to record the time at which events are ingested in order to track
latency as each travels through the topology. This is currently recorded for log
events using the Vector namespace in the `vector.ingest_timestamp` metadata
field, but we want it to be usable for all event types.  As such, we need a new
field in `struct EventMetadata`.

This change adds the new field as an `Option` so as to retain sane semantics for
`Default` implementations and to avoid extra calls to `Utc::now`.  The
`SourceSender::send` method sets this if the source doesn't so as to ensure
complete coverage.

For backward compatibility, this metadata is still inserted into the Vector log
namespace metadata, taken from this new field. Since this metadata is set up
before passing the events to the `SourceSender`, the ingest timestamp is set
manually in sources that can create Vector namespace logs.